+++ app/api/register/route.ts
@@ -1,8 +1,8 @@
-import { validate } from "@/app/middlewares/validate.middleware";
-import { prisma } from "@/app/utils/prisma";
+import { validate } from "@/middlewares/validate.middleware";
 import { SignUpInput, SignupSchema } from "@/app/validators/auth.validator";
 import { hash } from "bcryptjs";
 import { NextRequest, NextResponse } from "next/server";
+import { prisma } from "@/lib/prisma";
 
 const postHandler = async (
   req: NextRequest,
